[refactor/#71] useStore 파일명에서 use제거
Store라는 의미를 더 명시적으로 드러내기 위해 파일명에서 use제거
